ml-service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pandas
scikit-learn
joblib
python-dotenv
